+++ src/common/tools.c
@@ -4,7 +4,7 @@
 #include <stdbool.h>
 #include <stdlib.h>
 
-int euclidean_distance(int x1, int y1, int x2, int y2) {
+double euclidean_distance(int x1, int y1, int x2, int y2) {
     return sqrt((x1 - x2) * (x1 - x2) + (y1 - y2) * (y1 - y2));
 }
